give up with vite.config.ts for now
